chore: make e2e work again
related to OPT-333

+++ start-e2e.sh
@@ -3,11 +3,33 @@ echo "Starting E2E tests"
 sh ./start-selenium.sh &
 
 mvn clean package -Pproduction -DskipTests
-mkdir ./distro/target/distro
-tar -xzvf ./distro/target/*.tar.gz -C ./distro/target/distro
-./distro/target/distro/start-optimize.sh &
 
 RETRIES=6
+SLEEP_TIME=10
+URL="http://localhost:8080"
+COMMAND="curl -XGET $URL"
+
+node ./client/scripts/engine.js start 8080 &
+
+until ${COMMAND} > /dev/null;
+  do
+    sleep ${SLEEP_TIME}
+    RETRIES=$(( RETRIES - 1 ))
+  if [ $RETRIES = 0 ]; then
+    echo "Error: engine did not start!"
+    exit 1;
+  else
+    echo "Polling engine ... $RETRIES left"
+  fi
+done
+
+mkdir ./distro/target/distro
+tar -xzvf ./distro/target/*.tar.gz -C ./distro/target/distro
+
+echo "********STARTING OPTIMIZE********"
+./distro/target/distro/start-optimize.sh &
+
+RETRIES=10
 SLEEP_TIME=10
 URL="http://localhost:8090"
 COMMAND="curl -XGET $URL"
@@ -24,8 +46,10 @@ until ${COMMAND} > /dev/null;
   fi
 done
 
-node ./client/scripts/add_demo_data.js
 ./client/node_modules/.bin/wdio ./client/wdio-ci.conf.js
-pkill -f optimize-backend
+if [ $? -ne 0 ]; then exit 1; fi
 
-echo "Stopping E2E test"
+pkill -f optimize-backend
+pkill -f tomcat
+pkill -f elasticsearch
+pkill -f selenium-standalone
